+++ src/Ext/TechnoType/Body.h
@@ -251,8 +251,6 @@ public:
 		Valueable<ColorStruct> SelectedInfo_BelowColor;
 		Valueable<TechnoTypeClass*> FakeOf;
 		CustomPalette CameoPal;
-		Nullable<bool> NoRearmInEMPState;
-		Nullable<bool> NoRearmInTemporal;
 		Valueable<int> RateDown_Ammo;
 		Valueable<int> RateDown_Delay;
 		Valueable<int> RateDown_Cover;
@@ -262,6 +260,10 @@ public:
 		Nullable<bool> NoQueueUpToEnter;
 		Nullable<bool> NoQueueUpToUnload;
 		Valueable<bool> CanManualReload;
+		Nullable<bool> NoRearm_Inactive;
+		Nullable<bool> NoRearm_Temporal;
+		Nullable<bool> NoReload_Inactive;
+		Nullable<bool> NoReload_Temporal;
 
 		Nullable<AnimTypeClass*> Wake;
 		Nullable<AnimTypeClass*> Wake_Grapple;
@@ -511,8 +513,6 @@ public:
 			, SelectedInfo_BelowColor { { 0, 0, 0 } }
 			, FakeOf {}
 			, CameoPal {}
-			, NoRearmInEMPState {}
-			, NoRearmInTemporal {}
 			, RateDown_Ammo { -1 }
 			, RateDown_Delay { 0 }
 			, RateDown_Cover { 0 }
@@ -522,6 +522,10 @@ public:
 			, NoQueueUpToEnter {}
 			, NoQueueUpToUnload {}
 			, CanManualReload { false }
+			, NoRearm_Inactive {}
+			, NoRearm_Temporal {}
+			, NoReload_Inactive {}
+			, NoReload_Temporal {}
 
 			, Wake { }
 			, Wake_Grapple { }